fix(sql): avoid unexpected errors when over() clause for analytic function is missing
