+++ core/src/main/java/io/questdb/griffin/SqlExecutionContext.java
@@ -91,6 +91,8 @@ public interface SqlExecutionContext extends Closeable {
             boolean baseSupportsRandomAccess
     );
 
+    void clearAnalyticContext();
+
     void initNow();
 
     long getNow();
