+++ core/src/test/java/io/questdb/cutlass/text/SqlExecutionContextStub.java
@@ -117,6 +117,10 @@ public class SqlExecutionContextStub implements SqlExecutionContext {
     }
 
     @Override
+    public void clearAnalyticContext() {
+    }
+
+    @Override
     public void initNow() {
     }
