+++ core/src/test/java/io/questdb/griffin/engine/analytic/AnalyticFunctionTest.java
@@ -1,0 +1,525 @@
+/*******************************************************************************
+ *     ___                  _   ____  ____
+ *    / _ \ _   _  ___  ___| |_|  _ \| __ )
+ *   | | | | | | |/ _ \/ __| __| | | |  _ \
+ *   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+ *    \__\_\\__,_|\___||___/\__|____/|____/
+ *
+ *  Copyright (c) 2014-2019 Appsicle
+ *  Copyright (c) 2019-2022 QuestDB
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ ******************************************************************************/
+
+package io.questdb.griffin.engine.analytic;
+
+import io.questdb.griffin.AbstractGriffinTest;
+import org.junit.Test;
+
+public class AnalyticFunctionTest extends AbstractGriffinTest {
+
+    @Test
+    public void testAnalyticFunctionWithPartitionAndOrderByNonSymbol() throws Exception {
+        assertQuery("row_number\tprice\tts\n" +
+                        "1\t42\t1970-01-01T00:00:00.000000Z\n" +
+                        "2\t42\t1970-01-02T03:46:40.000000Z\n" +
+                        "3\t42\t1970-01-03T07:33:20.000000Z\n" +
+                        "4\t42\t1970-01-04T11:20:00.000000Z\n" +
+                        "5\t42\t1970-01-05T15:06:40.000000Z\n" +
+                        "6\t42\t1970-01-06T18:53:20.000000Z\n" +
+                        "7\t42\t1970-01-07T22:40:00.000000Z\n" +
+                        "8\t42\t1970-01-09T02:26:40.000000Z\n" +
+                        "9\t42\t1970-01-10T06:13:20.000000Z\n" +
+                        "10\t42\t1970-01-11T10:00:00.000000Z\n",
+                "select row_number() over (partition by price order by ts), price, ts from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " 42 price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticFunctionWithPartitionAndOrderBySymbolNoWildcard() throws Exception {
+        assertQuery("row_number\n" +
+                        "3\n" +
+                        "6\n" +
+                        "2\n" +
+                        "1\n" +
+                        "5\n" +
+                        "1\n" +
+                        "4\n" +
+                        "3\n" +
+                        "2\n" +
+                        "1\n",
+                "select row_number() over (partition by symbol order by symbol) from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticFunctionWithPartitionAndOrderBySymbolWildcardFirst() throws Exception {
+        assertQuery("price\tsymbol\tts\trow_number\n" +
+                        "0.8043224099968393\tCC\t1970-01-01T00:00:00.000000Z\t3\n" +
+                        "0.2845577791213847\tBB\t1970-01-02T03:46:40.000000Z\t6\n" +
+                        "0.9344604857394011\tCC\t1970-01-03T07:33:20.000000Z\t2\n" +
+                        "0.7905675319675964\tAA\t1970-01-04T11:20:00.000000Z\t1\n" +
+                        "0.8899286912289663\tBB\t1970-01-05T15:06:40.000000Z\t5\n" +
+                        "0.11427984775756228\tCC\t1970-01-06T18:53:20.000000Z\t1\n" +
+                        "0.4217768841969397\tBB\t1970-01-07T22:40:00.000000Z\t4\n" +
+                        "0.7261136209823622\tBB\t1970-01-09T02:26:40.000000Z\t3\n" +
+                        "0.6693837147631712\tBB\t1970-01-10T06:13:20.000000Z\t2\n" +
+                        "0.8756771741121929\tBB\t1970-01-11T10:00:00.000000Z\t1\n",
+                "select *, row_number() over (partition by symbol order by symbol) from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticFunctionWithPartitionAndOrderBySymbolWildcardLast() throws Exception {
+        assertQuery("row_number\tprice\tsymbol\tts\n" +
+                        "3\t0.8043224099968393\tCC\t1970-01-01T00:00:00.000000Z\n" +
+                        "6\t0.2845577791213847\tBB\t1970-01-02T03:46:40.000000Z\n" +
+                        "2\t0.9344604857394011\tCC\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t0.7905675319675964\tAA\t1970-01-04T11:20:00.000000Z\n" +
+                        "5\t0.8899286912289663\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t0.11427984775756228\tCC\t1970-01-06T18:53:20.000000Z\n" +
+                        "4\t0.4217768841969397\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "3\t0.7261136209823622\tBB\t1970-01-09T02:26:40.000000Z\n" +
+                        "2\t0.6693837147631712\tBB\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t0.8756771741121929\tBB\t1970-01-11T10:00:00.000000Z\n",
+                "select row_number() over (partition by symbol order by symbol), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticFunctionWithFilter() throws Exception {
+        assertQuery("author\tsym\tcommits\trk\n" +
+                        "user2\tETH\t3\t2\n" +
+                        "user1\tETH\t3\t1\n",
+                "with active_devs as (" +
+                        "    select author, sym, count() as commits" +
+                        "    from dev_stats" +
+                        "    where author is not null and author != 'github-actions[bot]'" +
+                        "    order by commits desc" +
+                        "    limit 100" +
+                        "), active_ranked as (" +
+                        "    select author, sym, commits, row_number() over (partition by sym order by commits desc) as rk" +
+                        "    from active_devs" +
+                        ") select * from active_ranked where sym = 'ETH'",
+                "create table dev_stats as " +
+                        "(" +
+                        "select" +
+                        " rnd_symbol('ETH','BTC') sym," +
+                        " rnd_symbol('user1','user2') author," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionAndOrderByNonSymbol() throws Exception {
+        assertQuery("rank\tprice\tts\n" +
+                        "1\t42\t1970-01-01T00:00:00.000000Z\n" +
+                        "2\t42\t1970-01-02T03:46:40.000000Z\n" +
+                        "3\t42\t1970-01-03T07:33:20.000000Z\n" +
+                        "4\t42\t1970-01-04T11:20:00.000000Z\n" +
+                        "5\t42\t1970-01-05T15:06:40.000000Z\n" +
+                        "6\t42\t1970-01-06T18:53:20.000000Z\n" +
+                        "7\t42\t1970-01-07T22:40:00.000000Z\n" +
+                        "8\t42\t1970-01-09T02:26:40.000000Z\n" +
+                        "9\t42\t1970-01-10T06:13:20.000000Z\n" +
+                        "10\t42\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by price order by ts), price, ts from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " 42 price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionAndOrderBySymbolNoWildcard() throws Exception {
+        assertQuery("rank\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n" +
+                        "1\n",
+                "select rank() over (partition by symbol order by symbol) from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionAndOrderBySymbolWildcardFirst() throws Exception {
+        assertQuery("price\tsymbol\tts\trank\n" +
+                        "0.8043224099968393\tCC\t1970-01-01T00:00:00.000000Z\t1\n" +
+                        "0.2845577791213847\tBB\t1970-01-02T03:46:40.000000Z\t1\n" +
+                        "0.9344604857394011\tCC\t1970-01-03T07:33:20.000000Z\t1\n" +
+                        "0.7905675319675964\tAA\t1970-01-04T11:20:00.000000Z\t1\n" +
+                        "0.8899286912289663\tBB\t1970-01-05T15:06:40.000000Z\t1\n" +
+                        "0.11427984775756228\tCC\t1970-01-06T18:53:20.000000Z\t1\n" +
+                        "0.4217768841969397\tBB\t1970-01-07T22:40:00.000000Z\t1\n" +
+                        "0.7261136209823622\tBB\t1970-01-09T02:26:40.000000Z\t1\n" +
+                        "0.6693837147631712\tBB\t1970-01-10T06:13:20.000000Z\t1\n" +
+                        "0.8756771741121929\tBB\t1970-01-11T10:00:00.000000Z\t1\n",
+                "select *, rank() over (partition by symbol order by symbol) from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionAndOrderBySymbolWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "1\t0.8043224099968393\tCC\t1970-01-01T00:00:00.000000Z\n" +
+                        "1\t0.2845577791213847\tBB\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t0.9344604857394011\tCC\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t0.7905675319675964\tAA\t1970-01-04T11:20:00.000000Z\n" +
+                        "1\t0.8899286912289663\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t0.11427984775756228\tCC\t1970-01-06T18:53:20.000000Z\n" +
+                        "1\t0.4217768841969397\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "1\t0.7261136209823622\tBB\t1970-01-09T02:26:40.000000Z\n" +
+                        "1\t0.6693837147631712\tBB\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t0.8756771741121929\tBB\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by symbol order by symbol), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionBySymbolAndOrderByPriceWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "2\t0.8043224099968393\tCC\t1970-01-01T00:00:00.000000Z\n" +
+                        "1\t0.2845577791213847\tBB\t1970-01-02T03:46:40.000000Z\n" +
+                        "3\t0.9344604857394011\tCC\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t0.7905675319675964\tAA\t1970-01-04T11:20:00.000000Z\n" +
+                        "6\t0.8899286912289663\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t0.11427984775756228\tCC\t1970-01-06T18:53:20.000000Z\n" +
+                        "2\t0.4217768841969397\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "4\t0.7261136209823622\tBB\t1970-01-09T02:26:40.000000Z\n" +
+                        "3\t0.6693837147631712\tBB\t1970-01-10T06:13:20.000000Z\n" +
+                        "5\t0.8756771741121929\tBB\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by symbol order by price), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_double(42) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionBySymbolAndOrderByIntPriceWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "1\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "4\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "4\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "1\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "1\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "1\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by symbol order by price), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionBySymbolAndOrderByIntPriceDescWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "2\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "1\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "2\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "1\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "2\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "2\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "2\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "2\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by symbol order by price desc), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionBySymbolAndNoOrderWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "1\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "1\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "2\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "2\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "3\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "4\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "3\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "4\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "2\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by symbol), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithPartitionBySymbolAndMultiOrderWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "1\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "4\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "4\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "1\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "1\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "1\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (partition by symbol order by symbol, price), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithNoPartitionByAndOrderBySymbolWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "3\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "7\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "7\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "3\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "3\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "3\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "7\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "7\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (order by symbol), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionWithNoPartitionByAndNoOrderByWildcardLast() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "1\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "1\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "1\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "1\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "1\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "1\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank() over (), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                null,
+                true,
+                true,
+                false
+        );
+    }
+
+    @Test
+    public void testAnalyticRankFunctionNonAnalyticContext() throws Exception {
+        assertQuery("rank\tprice\tsymbol\tts\n" +
+                        "1\t1\tBB\t1970-01-01T00:00:00.000000Z\n" +
+                        "1\t2\tCC\t1970-01-02T03:46:40.000000Z\n" +
+                        "1\t2\tAA\t1970-01-03T07:33:20.000000Z\n" +
+                        "1\t1\tCC\t1970-01-04T11:20:00.000000Z\n" +
+                        "1\t2\tBB\t1970-01-05T15:06:40.000000Z\n" +
+                        "1\t1\tBB\t1970-01-06T18:53:20.000000Z\n" +
+                        "1\t1\tBB\t1970-01-07T22:40:00.000000Z\n" +
+                        "1\t1\tCC\t1970-01-09T02:26:40.000000Z\n" +
+                        "1\t1\tCC\t1970-01-10T06:13:20.000000Z\n" +
+                        "1\t2\tAA\t1970-01-11T10:00:00.000000Z\n",
+                "select rank(), * from trades",
+                "create table trades as " +
+                        "(" +
+                        "select" +
+                        " rnd_int(1,2,3) price," +
+                        " rnd_symbol('AA','BB','CC') symbol," +
+                        " timestamp_sequence(0, 100000000000) ts" +
+                        " from long_sequence(10)" +
+                        ") timestamp(ts) partition by day",
+                "ts",
+                true,
+                true,
+                true
+        );
+    }
+}
